Extract interface BeaconTupleStorage
