+++ chain/src/main/java/org/ethereum/beacon/chain/storage/BeaconBlockStorage.java
@@ -2,8 +2,9 @@ package org.ethereum.beacon.chain.storage;
 
 import java.util.Optional;
 import org.ethereum.beacon.core.BeaconBlock;
+import tech.pegasys.artemis.ethereum.core.Hash32;
 
-public interface BeaconBlockStorage extends Hash32KeyStorage<BeaconBlock> {
+public interface BeaconBlockStorage extends HashKeyStorage<Hash32, BeaconBlock> {
 
   Optional<BeaconBlock> getCanonicalHead();
