+++ chain/src/main/java/org/ethereum/beacon/chain/storage/BeaconStateStorage.java
@@ -1,7 +1,8 @@
 package org.ethereum.beacon.chain.storage;
 
 import org.ethereum.beacon.core.BeaconState;
+import tech.pegasys.artemis.ethereum.core.Hash32;
 
-public interface BeaconStateStorage extends Hash32KeyStorage<BeaconState> {
+public interface BeaconStateStorage extends HashKeyStorage<Hash32, BeaconState> {
 
 }
