+++ chain/src/main/java/org/ethereum/beacon/chain/storage/HashKeyStorage.java
@@ -1,12 +1,11 @@
 package org.ethereum.beacon.chain.storage;
 
-import java.util.Optional;
 import org.ethereum.beacon.core.Hashable;
 import org.ethereum.beacon.db.source.DataSource;
-import tech.pegasys.artemis.ethereum.core.Hash32;
+import tech.pegasys.artemis.ethereum.core.Hash;
 
-public interface Hash32KeyStorage<T extends Hashable>
-    extends DataSource<Hash32, T> {
+public interface HashKeyStorage<H extends Hash, T extends Hashable<H>>
+    extends DataSource<H, T> {
 
   default void put(T item) {
     put(item.getHash(), item);
